requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
requests>=2.26.0,<3.0.0
python-dotenv>=0.19.0,<0.20.0
pydantic>=1.8.0,<2.0.0
aiofiles>=0.7.0,<0.8.0
python-slugify>=5.0.2,<6.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
httpx>=0.23.0
backoff>=2.0.0 